requirements.txt
# Core
asgiref==3.11.1
Django==5.2.11
sqlparse==0.5.5

# Database - PostgreSQL (production / optionnel en dev)
#psycopg2-binary==2.9.10

# Images (modèle Photo - ImageField)
Pillow==11.0.0

# Zone.boundary → surface_m2 (calcul géométrique, sans GDAL)
shapely>=2.0.0
pyproj>=3.6.0

# Export PDF
reportlab==4.2.5

# Configuration via variables d'environnement
django-environ==0.11.2

# Import Hydro-Québec (API)
requests==2.32.3
certifi>=2024.2.2  # Certificats SSL (évite erreurs sur macOS)

# API REST (mobile app)
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# Admin - menu personnalisé et regroupement (compatible Django 5)
django-modeladmin-reorder-reborn==0.1.4

# Serveur dev HTTPS (vue 3D Cesium avec imagerie Ion)
django-extensions==3.2.3
pyOpenSSL>=24.0.0
Werkzeug>=3.0.0
